requirements.txt
pandas>=2.2
numpy>=1.26
pyyaml>=6.0
matplotlib>=3.8
seaborn>=0.13
scikit-learn>=1.4
jupyter>=1.0
